+++ packages/web3-core/src/web3_subscriptions.ts
@@ -37,12 +37,23 @@ import { Web3SubscriptionManager } from './web3_subscription_manager.js';
 import { Web3EventEmitter, Web3EventMap } from './web3_event_emitter.js';
 import { Web3RequestManager } from './web3_request_manager.js';
 
+type CommonSubscriptionEvents = {
+	data: unknown; // Fires on each incoming block header.
+	error: Error; // Fires when an error in the subscription occurs.
+	connected: string; // Fires once after the subscription successfully connected. Returns the subscription id.
+};
+
 export abstract class Web3Subscription<
 	EventMap extends Web3EventMap,
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	ArgsType = any,
 	API extends Web3APISpec = EthExecutionAPI,
-> extends Web3EventEmitter<EventMap> {
+	// The following generic type is just to define the type `CombinedEventMap` and use it inside the class
+	// 	it combines the user passed `EventMap` with the `CommonSubscriptionEvents`
+	//	However, this type definition could be refactored depending on the closure of
+	//	[Permit type alias declarations inside a class](https://github.com/microsoft/TypeScript/issues/7061)
+	CombinedEventMap extends CommonSubscriptionEvents = EventMap & CommonSubscriptionEvents,
+> extends Web3EventEmitter<CombinedEventMap> {
 	private readonly _subscriptionManager: Web3SubscriptionManager<API>;
 	private readonly _lastBlock?: BlockOutput;
 	private readonly _returnFormat: DataFormat;
@@ -127,6 +138,8 @@ export abstract class Web3Subscription<
 			method: 'eth_subscribe',
 			params: this._buildSubscriptionParams(),
 		});
+
+		this.emit('connected', this._id);
 		return this._id;
 	}
 
@@ -154,14 +167,17 @@ export abstract class Web3Subscription<
 		this._id = undefined;
 	}
 
-	// eslint-disable-next-line class-methods-use-this, @typescript-eslint/no-unused-vars
-	protected _processSubscriptionResult(_data: unknown) {
-		// Do nothing - This should be overridden in subclass.
+	// eslint-disable-next-line class-methods-use-this
+	protected formatSubscriptionResult(data: CombinedEventMap['data']) {
+		return data;
 	}
 
-	// eslint-disable-next-line class-methods-use-this, @typescript-eslint/no-unused-vars
-	protected _processSubscriptionError(_err: Error) {
-		// Do nothing - This should be overridden in subclass.
+	public _processSubscriptionResult(data: CombinedEventMap['data'] | unknown) {
+		this.emit('data', this.formatSubscriptionResult(data));
+	}
+
+	public _processSubscriptionError(error: Error) {
+		this.emit('error', error);
 	}
 
 	// eslint-disable-next-line class-methods-use-this