+++ packages/web3-eth/CHANGELOG.md
@@ -155,6 +155,12 @@ Documentation:
 
 -   Dependencies updated
 
+## [4.0.3]
+
+### Changed
+
+-   Dependencies updated
+
 ## [Unreleased]
 
 ### Fixed