+++ packages/web3-types/CHANGELOG.md
@@ -134,8 +134,10 @@ Documentation:
 
 -   Fixed bug #6185, now web3.js compiles on typescript v5 (#6195)
 
-## [Unreleased]
+## [1.0.2]
 
 ### Fixed
 
 -   type `Filter` includes `blockHash` (#6206)
+
+## [Unreleased]